fix: revert resend job
